prettier applied
SWAL alerts added for modals.

refresh upon completion of upload or invite sent
