+++ cypress/plugins/index.js
@@ -1,5 +1,5 @@
 // ***********************************************************
-// This example plugins/UploadPDF.jss can be used to load plugins
+// This example plugins/InviteModala.jss can be used to load plugins
 //
 // You can change the location of this file or turn off loading
 // the plugins file with the 'pluginsFile' configuration option.
