+++ cypress/support/index.js
@@ -1,5 +1,5 @@
 // ***********************************************************
-// This example support/UploadPDF.jss is processed and
+// This example support/InviteModala.jss is processed and
 // loaded automatically before your test files.
 //
 // This is a great place to put global configuration and
