+++ package.json
@@ -27,7 +27,8 @@
     "redux": "^4.0.5",
     "redux-thunk": "^2.3.0",
     "segmented-control": "^0.1.12",
-    "styled-components": "^5.0.0"
+    "styled-components": "^5.0.0",
+    "sweetalert": "^2.1.2"
   },
   "scripts": {
     "start": "react-scripts start",
